Exam-Budget-Tracker-App/server/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
pytest==7.4.4
pytest-django==4.7.0
gunicorn==21.2.0
